new page navigation for students
